wip: component item rendering | image colouring of attributes ...

+++ gui/NMLayer.h
@@ -161,7 +161,7 @@ signals:
 	void layerProcessingStart();
 	void layerProcessingEnd();
 	void layerLoaded();
-	void notifyLastClickedRow(NMLayer* l, long cellID);
+	void notifyLastClickedRow(NMLayer* l, double cellID);
 
 protected:
 	vtkSmartPointer<vtkRenderWindow> mRenderWindow;
